Move logic to a private internal/ folder
Reduces public footprint
