+++ errors.go
@@ -2,8 +2,9 @@ package sipuri
 
 import (
 	"errors"
-	"strconv"
 	"strings"
+
+	"github.com/percivalalb/sipuri/internal"
 )
 
 // ErrInvalidScheme is returned when a string that does not start sip: or sips: is given.
@@ -88,16 +89,4 @@ func (err MalformedURIError) Unwrap() error {
 }
 
 // EscapeError is returned when a byte-pair has been incorrectly URL encoded.
-type EscapeError string
-
-// Error returns the string representation of the error.
-func (e EscapeError) Error() string {
-	return "sip: invalid URL escape " + strconv.Quote(string(e))
-}
-
-// Is allows [EscapeError] to be compared by [errors.Is].
-func (e EscapeError) Is(input error) bool {
-	_, ok := input.(EscapeError)
-
-	return ok
-}
+type EscapeError = internal.EscapeError
